Add uncapturederror test
It wasn't clear if I should just put this test in error_scope.spec.ts
or put it here where there was a TODO but technically this isn't
error scope so ?
