+++ src/webgpu/listing_meta.json
@@ -212,7 +212,6 @@
   "webgpu:api,operation,texture_view,write:aspect:*": { "subcaseMS": 0.700 },
   "webgpu:api,operation,texture_view,write:dimension:*": { "subcaseMS": 0.601 },
   "webgpu:api,operation,texture_view,write:format:*": { "subcaseMS": 0.600 },
-  "webgpu:api,operation,uncapturederror:constructor:*": { "subcaseMS": 0.200 },
   "webgpu:api,operation,uncapturederror:iff_uncaptured:*": { "subcaseMS": 0.101 },
   "webgpu:api,operation,uncapturederror:only_original_device_is_event_target:*": { "subcaseMS": 0.101 },
   "webgpu:api,operation,uncapturederror:uncapturederror_from_non_originating_thread:*": { "subcaseMS": 0.201 },
